US-Backend/requirements.txt
Flask
flask-cors
openai
python-dotenv
fastapi
uvicorn
motor
pymongo